New planetary system generation with new planet types | Add new Ui widgets for layer design | Bug Fixes

+++ Space Game/Core/GameObjects/GameObject.cs
@@ -17,7 +17,7 @@ namespace Galaxy_Explovive.Core.GameObject
         public float TextureSclae { get; set; }
         public int TextureWidth { get; set; }
         public int TextureHeight { get; set; }
-        public float TextureDepth { get; set; }
+        public int TextureDepth { get; set; }
         public Color TextureColor { get; set; }
         public CircleF BoundedBox { get; set; }
